feat: replace `<OAuth>` with sync button
